move introspection logic, reformat tests
