+++ dcpy/lifecycle/ingest/transform.py
@@ -7,7 +7,7 @@ from typing import Callable, Literal
 
 from dcpy.models import file
 from dcpy.models.lifecycle.ingest import FunctionCall
-from dcpy.utils import data, validation
+from dcpy.utils import data, introspect
 from dcpy.utils.logging import logger
 from dcpy.connectors.edm import recipes
 
@@ -205,7 +205,7 @@ def validate_pd_series_func(
             return f"'{func_str}' has no attribute '{part}'"
         func = func.__getattribute__(part)
         func_str += f".{part}"
-    return validation.validate_function_args(func, kwargs)  # type: ignore
+    return introspect.validate_function_args(func, kwargs)  # type: ignore
 
 
 def validate_processing_steps(
@@ -228,7 +228,7 @@ def validate_processing_steps(
             func = getattr(preprocessor, step.name)
 
             # assume that function takes args "self, df"
-            kw_error = validation.validate_function_args(
+            kw_error = introspect.validate_function_args(
                 func, step.args, raise_error=False, ignore_args=["self", "df"]
             )
             if kw_error:
